+++ src/main/java/umc/teumteum/server/domain/home/controller/HomeController.java
@@ -6,7 +6,10 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import umc.teumteum.server.domain.home.dto.*;
+import umc.teumteum.server.domain.home.dto.request.TodoRequestDTO;
+import umc.teumteum.server.domain.home.dto.response.TodoIdResponseDTO;
+import umc.teumteum.server.domain.home.dto.response.TodoInfoResponseDTO;
+import umc.teumteum.server.domain.home.dto.response.TodyTeumResponseDTO;
 import umc.teumteum.server.domain.home.exception.status.HomeSuccessStatus;
 import umc.teumteum.server.domain.home.service.HomeService;
 import umc.teumteum.server.global.apiPayload.ApiResponse;
