+++ src/main/java/umc/teumteum/server/domain/home/controller/WishController.java
@@ -6,12 +6,12 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import umc.teumteum.server.domain.home.dto.WishDeleteRequestDTO;
-import umc.teumteum.server.domain.home.dto.WishInfoResponseDTO;
-import umc.teumteum.server.domain.home.dto.WishRequestDTO;
+import umc.teumteum.server.domain.home.dto.request.WishDeleteRequestDTO;
+import umc.teumteum.server.domain.home.dto.response.WishInfoResponseDTO;
+import umc.teumteum.server.domain.home.dto.request.WishRequestDTO;
 import umc.teumteum.server.domain.home.exception.status.HomeSuccessStatus;
 import umc.teumteum.server.domain.home.service.HomeService;
-import umc.teumteum.server.domain.home.dto.WishlistResponseDTO;
+import umc.teumteum.server.domain.home.dto.response.WishlistResponseDTO;
 import umc.teumteum.server.domain.user.entity.User;
 import umc.teumteum.server.global.annotation.CurrentUser;
 import umc.teumteum.server.global.apiPayload.ApiResponse;
