--- src/main/java/umc/teumteum/server/domain/home/dto/TodyTeumResponseDTO.java
@@ -1,4 +1,0 @@
-package umc.teumteum.server.domain.home.dto;
-
-public class TodyTeumResponseDTO {
-}
