+++ src/main/java/umc/teumteum/server/domain/home/dto/response/TodoIdResponseDTO.java
@@ -1,4 +1,4 @@
-package umc.teumteum.server.domain.home.dto;
+package umc.teumteum.server.domain.home.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
