+++ src/main/java/umc/teumteum/server/domain/home/dto/response/TodyTeumResponseDTO.java
@@ -1,0 +1,4 @@
+package umc.teumteum.server.domain.home.dto.response;
+
+public class TodyTeumResponseDTO {
+}
